requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
langchain==0.0.335
langchain-google-genai==0.0.6
google-generativeai==0.3.1
pypdf==3.17.1
faiss-cpu==1.8.0
python-dotenv==1.0.0
pydantic==2.4.2
numpy==1.26.2
sentence-transformers==2.2.2